python/insta_agent/requirements.txt
agent-framework[azure,devui]==1.0.0b260130
azure-identity>=1.19.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
azure-monitor-opentelemetry>=1.0.0
azure-storage-blob>=12.20.0
azure-cosmos>=4.9.0
azure-communication-email>=1.0.0
Pillow>=10.0.0
fal-client>=0.5.0
uvicorn>=0.27.0
